Readability and cleanup
