--- docker_entry.sh
@@ -1,2 +1,0 @@
-/app/pagemail/server --dir=/data/pb_data --http=0.0.0.0:4000 serve &
-nginx -g "daemon off;"
