--- polly_output.json
@@ -1,25 +1,0 @@
-{
-  "ResponseMetadata": {
-    "RequestId": "6b17d1a6-def2-4f22-9d64-0e71101f8c13",
-    "HTTPStatusCode": 200,
-    "HTTPHeaders": {
-      "x-amzn-requestid": "6b17d1a6-def2-4f22-9d64-0e71101f8c13",
-      "content-type": "application/json",
-      "content-length": "472",
-      "date": "Thu, 27 Jul 2023 23:48:54 GMT"
-    },
-    "RetryAttempts": 0
-  },
-  "SynthesisTask": {
-    "Engine": "standard",
-    "TaskId": "2ce9ece0-128f-41bb-8d40-7e723f405d5b",
-    "TaskStatus": "scheduled",
-    "OutputUri": "https://s3.eu-west-2.amazonaws.com/pagemail-speechsynthesis/test2.2ce9ece0-128f-41bb-8d40-7e723f405d5b.mp3",
-    "CreationTime": "2023-07-28 00:48:54.636000+01:00",
-    "RequestCharacters": 6501,
-    "OutputFormat": "mp3",
-    "TextType": "ssml",
-    "VoiceId": "Amy",
-    "LanguageCode": "en-GB"
-  }
-}
