+++ readability/parser/mock.py
@@ -1,0 +1,3 @@
+from unittest.mock import Mock
+
+session = Mock()
