+++ test_pages/readable.html
@@ -93,9 +93,7 @@
 					<span class="astro-IS5RL2GN"
 						style="--onDark0: #fa5252;--onDark1: #fab005;--onDark2: #40c057;--onDark3: #228be6;--onDark4: #be4bdb;--onLight0: #e03131;--onLight1: #f59f00;--onLight2: #2f9e44;--onLight3: #1971c2;--onLight4: #9c36b5;">CPU</span></span>
 			</h1>
-			<p class="abstract">Curious exactly what happens when you run a program on your computer? Read this article
-				to learn how multiprocessing works, what system calls really are, how computers manage memory with
-				hardware interrupts, and how Linux loads executables.</p>
+			<p class="abstract">Curious exactly what happens when you run a program on your computer? </p>
 			<p>
 				By
 				<a target="_blank" href="https://kognise.dev/">Lexi Mattick</a>
@@ -210,16 +208,13 @@
 			<p>I’ve done <a href="https://github.com/kognise" rel="noopener noreferrer" target="_blank">a lot of things
 					with computers</a>, but I’ve always had a gap in my knowledge: what exactly happens when you run a
 				program on your computer?
-			<p>And you know what they say… you only truly understand something if you can explain it to someone else.
 			</p>
 			<blockquote>
 				<p>In a hurry? Feel like you know this stuff already?</p>
-				<p><a href="/how-to-run-a-program">Read chapter 3</a> and I guarantee you will learn something new.
-					Unless you’re like, Linus Torvalds himself.</p>
 			</blockquote>
 
 			<a class="continue" href="the-basics">
-				Continue to Chapter 1: The “Basics”
+				Continue.
 				<span aria-hidden="true">&raquo;</span>
 			</a>
 		</div>
